Imu fall detection working

+++ components/smart_signage/include/imu/imu_fsm.h
@@ -68,7 +68,7 @@ class FSM {
     void onFallenExit();
     void onError();
 
-    double computeTiltAngle(const Vector &curAccel, const Vector &refAccel) const;
+    uint16_t computeTiltAngle(const Vector &curAccel, const Vector &refAccel) const;
 
     ctrl::Q       &ctrlQ_;
     hal::IImuHal  &hal_;
